Create GlTexture class
